Moved shared elements into separate folder
