+++ Examples/SharedElements/index.js
@@ -1,77 +1,13 @@
 import React from 'react';
-import { View, Text, ScrollView, Image, Dimensions, TouchableOpacity, StyleSheet } from 'react-native';
+import { View, Text, ScrollView, Image, Dimensions, TouchableOpacity } from 'react-native';
 import { withNavigation } from 'react-navigation';
 import Icon from 'react-native-vector-icons/MaterialCommunityIcons';
-import { Transition, createFluidNavigator } from '../lib';
+import { Transition, createFluidNavigator } from '../../lib';
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-  card: {
-    backgroundColor: '#FFF',
-    flexDirection: 'column',
-    justifyContent: 'center',
-    margin: 20,
-    marginTop: 10,
-    marginBottom: 10,
-    borderRadius: 4,
-    shadowOpacity: 0.3,
-    shadowColor: '#AAA',
-    shadowOffset: { width: 0, height: 4 },
-  },
-  smallImage: {
-    width: Dimensions.get('window').width - 40,
-    height: Dimensions.get('window').width - 60,
-  },
-  imageHeader: {
-    padding: 10,
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-  avatarImage: {
-    width: 20,
-    height: 20,
-    borderRadius: 10,
-  },
-  avatarSmallImage: {
-    width: 12,
-    height: 12,
-    borderRadius: 6,
-  },
-  avatarText: {
-    fontSize: 11,
-    marginLeft: 8,
-  },
-  bigCard: {
-    ...StyleSheet.absoluteFill,
-    backgroundColor: '#FFF',
-    justifyContent: 'flex-start',
-  },
-  bigImage: {
-    width: Dimensions.get('window').width,
-    height: Dimensions.get('window').width,
-  },
-  commentsContainer: {
-  },
-  commentRow: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    padding: 10,
-    paddingVertical: 6,
-  },
-  comment: {
-    fontSize: 11,
-    paddingLeft: 6,
-  },
-  buttons: {
-    flexDirection: 'row',
-    padding: 10,
-    paddingTop: 50,
-  },
-});
+import styles from './styles';
+import { createItemsData } from './data';
 
-const ImageHeader = () => (
+const ImageFooter = () => (
   <View style={styles.imageHeader}>
     <Icon name="heart-outline" size={22} />
     <View style={{ width: 10 }} />
@@ -111,7 +47,7 @@ const Card = ({ navigation, item, id }) => (
         <Image style={styles.smallImage} source={item.source} />
       </Transition>
       <Transition shared={`imageHeader${id}`}>
-        <ImageHeader />
+        <ImageFooter />
       </Transition>
     </TouchableOpacity>
   </Transition>
@@ -162,9 +98,9 @@ const Screen2 = ({ navigation }) => (
           <Image style={styles.bigImage} source={navigation.getParam('item').source} />
         </Transition>
         <Transition shared={`imageHeader${navigation.getParam('id')}`}>
-          <ImageHeader />
+          <ImageFooter />
         </Transition>
-        <Transition anchor={`image${navigation.getParam('id')}`} appear="horizontal">
+        <Transition anchor={`card${navigation.getParam('id')}`}>
           <ScrollView style={styles.commentsContainer}>
             {navigation.getParam('item').comments.map((comment, index) => (
               <Comment key={index} comment={comment} />
@@ -193,7 +129,6 @@ const Navigator = createFluidNavigator({
   navigationOptions: {
     gesturesEnabled: true,
     gestureResponseDistance: {
-      horizontal: Dimensions.get('window').width,
       vertical: Dimensions.get('window').height,
     },
   },
@@ -209,48 +144,5 @@ class SharedElements extends React.Component {
     );
   }
 }
-
-const createItemsData = () => {
-  const users = getRandomImages(15, 30).map(img => ({
-    source: img,
-    name: ['Marge Manns',
-      'Wallace Doe',
-      'Cheryle Hodnett',
-      'Jared Muszynski',
-      'Jayme Poyer',
-      'Gina Dennett'][Math.floor((Math.random() * 6))],
-  }));
-
-  const createComments = () => {
-    const comments = [];
-    for (let i = 0; i < Math.floor((Math.random() * 6) + 4); i++) {
-      comments.push({
-        text: ['Great picture! Thanks', 'I like it!', 'Wow! Great', 'Cool!', 'Thanks :-)', ':-)'][Math.floor((Math.random() * 6))],
-        avatar: users[Math.floor((Math.random() * 15))],
-      });
-    }
-    return comments;
-  };
-  const items = getRandomImages(15, Dimensions.get('window').width).map(img => ({
-    source: img,
-    comments: createComments(),
-    avatar: users[Math.floor((Math.random() * 15))],
-  }));
-  return items;
-};
-const getRandomImages = (count: number, size: number) => {
-  const items = [];
-  const randMax = 100;
-  for (let i = 0; i < count; i++) {
-    let randomNumber = Math.floor((Math.random() * randMax) + 1);
-    const idExists = (e) => e.id === randomNumber;
-    while (items.findIndex(idExists) > -1) {
-      randomNumber = Math.floor((Math.random() * randMax) + 1);
-    }
-
-    items.push({ url: `https://picsum.photos/${size}/${size}?image=${randomNumber}`, id: randomNumber });
-  }
-  return items;
-};
 
 export default SharedElements;
